Add command mode
This mode is entered before serial line begins. Currently implemented
commands are: dtr, rst and go. "dtr" and "rst" are the same as -d and -r
options, "go" ends command mode and enters serial line interaction.

This is intended for use by novaboot in this way: Novaboot connects to
the serial line to be sure that the target is not occupied. Then it
copies the images, which can take long time, and after that it sends
reset command to reset the target. Without this, the target is reset (by
-d or -r) before copying the images and after that it may be too late to
communicate with the bootloader.