Handle hang-up event on stdin
When sterm is run with pipe as stdin rather than terminal, as is the
case, for example, with the following command:

     ssh somewhere sterm /dev/ttyS0

closing the pipe (e.g. killing ssh) does not send SIGHUP signal, but
POLLHUP event is reported by poll(). When this happened, sterm ended
up in an infinite loop, consuming 100% CPU. Now, we handle this event
and exit.